refactor(write-path): split CalculationJobService into External API + Write Path services
CalculationExecutionService handles Write Path (startCalculation,
handleCalculationFailure, completeCalculationWithResult).
CalculationJobService now only handles External API Path (createJob,
OCID resolve, snapshot ready, API/OCID failure retry).
Removed unused methods: requestApiData, completeCalculation.
